can: initial version using C bindings
This patch adds a version of the can plugin that uses the C bindings. It
is actually written in C++.  It tries to be as close as possible to the
Python version, so that we can compare both.

To decode message databases (dbc files), it uses the cantools library
available here:

    https://github.com/aheit/cantools

Signed-off-by: Simon Marchi <simon.marchi@efficios.com>