can/c: make bitfield.h C++-friendly
When compiling bitfield.h as C++, g++ doesn't like how we implement a
static assert:

  bitfield.h:74:16: error: types may not be defined in ‘sizeof’ expressions
  sizeof(struct { int f:(2 * !!(cond) - 1); })
                ^

Ifdef it for C++, and use static_assert (available starting with
C++11).  A little trick is needed to make it work in the context of
_BT_BUILD_ASSERT.  _BT_BUILD_ASSERT is used in the context where it
needs to yield a value (we don't care what it is, it just need to be
some integer value).  static_assert doesn't yield a value.  This is
solved by using a lambda expression with the static_assert inside, and
the comma operator, which returns the value on the right (a 0).

Another small error is fixed, which to to cast a pointer to the right
type instead of void *.  This should be valid for C as well.

Signed-off-by: Simon Marchi <simon.marchi@efficios.com>