can/graph/gui: Initial proof-of-concept of Babeltrace2 - Qt5 GUI application.
Babeltrace graph processing loop is run in a separate python thread.
This has the obvious drawback of having little control on thread
scheduling, which causes the graph thread to starve the gui thread.

The examples provide (not so pretty) workarounds to mitigate this.

can_graph_gui_simple.py is the graph_gui implementation with signals,
which is simpler to understand and might actually be good enough
if you process only a few events.

can_graph_gui_responsive.py refers to the EventBuffer implementation
with tweaks to make the GUI more responsive.
It also presents basic statistics + has an option to follow events.