can/python: Provided a demonstration for class field / payload field parsing.
The can_graph.py was extended with can_user_detail example, which
implements a sink with similar functionaility as text.detail sink.
It is capable of dumping event class information from the
'stream beginning' message.

The advanced gui example was extended with a QTreeView, which is
initially populated with event class / field class information,
(retrieved from 'stream beginning' message)
with a live view of event payload fields as the messages get
parsed by the sink.

We use QStandardItemModel for QTreeView, instead of QAbstractItemModel,
which is suggested by Qt documentation - the QStandardItemModel
provides content formatting / view update functionality.

Also removed EventBuffer since this "optimized" data structure
doesn't really make a considerable difference in terms of graph
execution speed - the major bottleneck
(that we're content living with for now) is the GUI.

I might ressurect it for a multiprocessing example later on.

The simplified, lazy loaded variant table model is now AppendableTableModel.